Hook `search-api-v2` up to shared Redis
Adds `REDIS_URL` env variable pointing to the shared Redis to the
`search-api-v2` app across environments for use in an upcoming
sync-related locking/mutex feature.

This isn't ideal, and in the future we will want to disentangle the
myriad use cases of that shared Redis instance, but this'll do for the
short term.

c.f. https://gds.slack.com/archives/C013F737737/p1709897264741419