Run integration ES restore outside off-peak window.
Still don't know for sure why the Elasticsearch restore cron is flaking
in the integration cluster (having worked reliably for months), so try a
couple more stabs in the dark to see if the problem goes away.

If it's still flakey after this then we'll take a step back and consider
alternatives (or maybe delve in to find out exactly why it's flaking,
but that's probably a big timesink for what's already a bit of a whacky
and non-essential cronjob).

- Wait longer between issuing the delete and issuing the restore. Just a
  hunch (e.g. maybe integration is slower than the other clusters for
  some reason?)
- Change the cron schedule to avoid the AWS maintenance window
  (20:00-06:00 UTC) for the integration ES cluster. Also just a hunch.