Suspend search-index-env-sync cron in integration.
Now that site search no longer uses Elasticsearch, there's little reason
to copy the Elasticsearch indices from staging to production every day.

The job has lately started flaking because (for as yet unknown reasons)
the search-api-worker deployment seems to be doing a lot of inserts
which re-create the index in between the `DELETE` and the `POST
_restore` (even if we issue them in immediate succession). This leaves
the integration environment in a non-working state every few days, which
is way worse than some data being a bit stale (that probably nobody even
needs anyway).

There are a few features that still use Elasticsearch, such as
[finders], but at this point it's likely that the disadvantages of
clobbering the integration indices every morning outweigh the
advantages.

It's easy enough to run the job on-demand (from the Argo CD web UI or
the command line) if it's needed — and I doubt it'll even be needed.

Also remove the delay between the delete and the restore, since we now
know this exacerbates the race condition rather than mitigating it.

[finders]: https://github.com/alphagov/finder-frontend#live-examples