Work around gotcha in appImage value.
Turns out the `appImage` Helm value is a bit of a confusing special
case cos it gets mangled through
app-config/templates/govuk-application.yaml.

So setting anything under `appImage` in `helmValues` in `values-*.yaml`
shadows the `appImage` value object that comes out of charts/app-config,
removing the rest of the info about the app image. Confusing.

Simpler to make `arch` a top-level value in charts/generic-govuk-app.
Also kinda makes sense cos `arch` is really about a
scheduling-constraint so affects the whole pod/deployment and not just
the app image (even though it is arguably a property of the app image).