Reinstate default replicaCount in prod and staging.
Looks like we accidentally these in d445927c :) No harm done.

This adds weight to the idea that we can probably drop to N+1 on a bunch
of things (basically anything that's both non-public-facing and
non-critical, i.e. not a problem if it's unavailable for a handful of
minutes a month).