Suspend database backup / restore in staging
This is to allow us to preview "history mode" ahead of the general
election. We're suspending restores so that we've got time to evaluate
the impact of history mode (i.e. which pages will it / won't it affect,
and will they look correct).

I only really need to suspend the restores, but the config doesn't
separate out the ops into separately suspendable tasks, so it's easier
to just suspend the backups too. Backups in staging are not really
important.