Perform environment sync operations at the same time
For Publishing API and Content Store, we need their databases to be in
sync with each other when the backup is taken and restored during the
environment sync.

If we don't, we can get data conflicts that stop content being published
to Content Store.

Therefore bringing backups for Publishing API and Content Store to be at
the same time, and the same for the restore jobs.