Add 'backup' operations for integration environment
What:

Ensures that every application has its database data backed up
to S3 in integration, for continued compatibility with
govuk-docker.

Why:

Since the env sync was ported over to govuk-helm-charts, not all
apps have backups saved to s3 on integration. This means that the
data replication scripts in govuk-docker fail, as they're looking
for a bucket that does not exist.

Whilst we could configure govuk-docker to pull from production or
staging instead (and indeed explored the idea in
https://github.com/alphagov/govuk-docker/pull/763), this has the
downside of requiring that engineers have
[Production Admin access](https://docs.publishing.service.gov.uk/manual/rules-for-getting-production-access.html#production-admin-access),
so new starters would be unable to easily get production-like
data for local development (new starters are only granted
Integration Admin access).