Add finder-frontend to draft stack
We're hoping to use the draft stack to support 'previews' of
specialist finders, to avoid publishers having to use Integration
for that.

This chart definition defines the same `rails`, `sentry` and
`uploadAssets` values as in the other `draft-*` charts, and
copies the `cronTasks` from the live stack definition.

Note that we don't have arm nodes in production or staging yet,
so they're deliberately omitted:
https://github.com/alphagov/govuk-helm-charts/pull/2237#pullrequestreview-2224531811

NB, I also noticed a few inconsistencies with other apps, which
have been escalated to the Patterns and Pages team:

- [Broken draft-email-alert-frontend](https://gds.slack.com/archives/C06T3EFUUQ6/p1722859514105799)
- [Incomplete info-frontend migration](https://gds.slack.com/archives/C06T3EFUUQ6/p1722859637714859)
- [Suggestion to add 'feedback' to draft stack for consistency](https://gds.slack.com/archives/C06T3EFUUQ6/p1722859913038219)

Trello: https://trello.com/c/cevf50zS/2451-spike-specialist-finder-previews-kick-off-needed