Promote GOV.UK Chat users in integration and staging
This adds the task we added for the production environment [1] to the
staging and integration environments as well. I've set this to not run
every hour here but just copying a precedent from a different one that
is 7am to 7pm - which seemed a good shout to avoid any data sync issues.

[1]: https://github.com/alphagov/govuk-helm-charts/pull/2343