Allow clamav to write log file on startup
As of ClamAV 1.3.2, log files no longer can reference symlinks which mean that it cannot be explicitly configured to /dev/stdout. However the default behaviour is to write to stdout anyways, but without explicit config it tries to touch log files under /var/logs/clamav. This gives it write permission to do so.