Set GOV.UK Chat places env vars to 0
If we merged this PR https://github.com/alphagov/govuk-helm-charts/pull/2450
we would immediately start incrmenting the number of sign up places
available in the app.

By setting the env vars for places to 0, we can ensure that sign ups
aren't incremented until we're ready.

At which point we will revert these values.