Set GOV.UK Chat sidekiq concurrency as env vars
We previously didn't specify this as env vars and just had a concurrency
set in the Sidekiq config file, however we had made a mistake that there
wasn't a large enough DB connection pool to satisfy that level of
concurrency.

As we don't want to change the size of DB connection pool everywhere, we
only set it as an env var for this process. Since we want it to match
the concurrency level of sidekiq it seems logical to have it defined
together.

This pairs with an application change in: https://github.com/alphagov/govuk-chat/pull/709