Update chat cron job schedules
This sets the waiting list promote job to run at half past the hour and
to therefore not clash with instant access places. In production we've
also limited it to run from 7am to 10pm so as to not send out emails
through the night about places - to try avoid a rush of users seeing the
email in the morning at a similar UK time.

It also nudges the delayed access to be run at quarter past to space it
out a bit between the jobs - though this probably doesn't really matter
much.