Add Slack alerts for GOV.UK Chat OpenAI rate limits
The commit adds two Slack alerts, one for request and one for token
rate limits. These are triggered if more than 80% of the limit is
used for any model at any time.

Alerts sent to Slack include a description with the model that
the alert was triggered for, and the endpoint that called it.