Add new Redis URLs for Emergency Banner and Taxonomy Cache
Whitehall uses Redis for three purposes:
- Taxonomy Cache
- Emergency Banner
- GOV.UK Sidekiq

We've created a new dedicated Redis instance for this to move it off the
shared instance, and using different numerically-named databases which
are accessed via the trailing `/1` or `/2` on that instance, rather than
namespacing.

Whitehall and Static need to share a Redis instance for the Emergency
banner, so we're storing this as a top level YAML anchor.