Remove expectation of Redis/Sidekiq from `search-admin`
The `search-admin` app doesn't have any workers or any other usage of
Redis, so it doesn't need to be configured as such.

Therefore removing the `REDIS_URL` environment variable and switching
off the expectation of workers.

This means the app does not need to be migrated to a Redis instance of
it's own.