Add internal ALB to production search-api-v2
Adding to production now. Staging and integration were done in https://github.com/alphagov/govuk-helm-charts/pull/2705

This is a new load balancer, because it behaves differently from the backend apps (this one is internal to the VPC), and will require different WAF settings because it will have a different traffic profile.

It allows traffic to be routed to search-api-v2 from applications outside the EKS cluster such as an API gateway.

We're doing this to allow publicly routed search API requests, while maintaining more control over what happens than usual. We don't want to adopt the previous approach of allowing open access to the API via a www.gov.uk route (eg www.gov.uk/api/search-v2.json) because of the cost implications. Because of that, this load balancer allows internal traffic only.