Change integration refresh from daily to weekly
GOV.UK's Integration environment is used by various groups of people for
various things:

- It's used by developers as the start of the deployment pipeline.
  Unstable code is deployed there regularly, often from branches which
may not have been reviewed or tested.
- It's used by developers to reproduce issues originally found in other
  environments, allowing them to experiment with solutions which would
be too risky to try in production (e.g. changing data using a console)
- It's used by content designers in GDS and other government departments
  as a training environment (Whitehall and other publishing apps such as
Specialist), both formally and informally
- It's data is extracted as the starting point for various ETL
  pipelines, for example govsearch
- It's used occasionally as a way to preview bits of content which are
  not editionable (and therefore can't be previewed on the draft stack)

Currently, data from GOV.UK's production databases are copied each night
from the production environment to the staging environment, and then
onwards to the integration environment.

Having up to date data on integration is useful for developers. Because
the data on integration is usually no more than 1 day behind production,
any bugs found on the production website are likely to be reproducible
in integration where it's safer to experiment with changes.

However, the daily sync also overwrites any content published on
integration itself, and this is not useful for content designers.
Training is made much more complicated because content has to be
drafted, reviewed and published all in the same day, and if there's a
delay they have to start again from square one the following day. If
something is being drafted on integration, it's only previewable for one
day before it's cleared down.

In the long term, we hope to separate out the content design use cases
from the developer use cases, and have a separate stable training
environment for the content designers. In the short term however,
reducing the frequency of these refreshes seems pragmatic.

Consequences of this change:

- Govsearch data (and other data ETL'd from integration) will be up to 1
  week stale, i.e. it will be missing newly published and updated
  documents
- If there are issues with newly published content which developers need
  to investigate, they'll have to use the staging environment instead of
  the integration environment.
- Content designers will have a much more pleasant experience doing
  training.
- Integration will be more useful for previewing non-editionable content
  (as it will survive more than one day)

Note I've chosen Monday morning as the day for the weekly refresh,
rather than Saturday or Sunday because staging only refreshes on week
days, so a refresh on a Sunday would miss anything published on Saturday
or Sunday (because it won't have made it to staging).