Add a global variable to enable search autocomplete
The new search autocomplete is due to be rolled out across GOV.UK

Autocomplete uses a denylist to stop certain terms from serving
suggestions. The process to update the denylist is very manual and
developer heavy.

If might be difficult to update the denylist over the holiday period.
In that situation the decision might be taken to temporarily turn off
autocomplete entirely.

A global environment variable has been added to each environment rather
than individual environment variables for each affect app, as the search
bar exists in the header on all pages on GOV.UK. This means that it
needs to be accessed by the [layout_super_navigation_header](https://github.com/alphagov/govuk_publishing_components/pull/4451/files) which is pulled in by
static for use in 6 other rendering apps.