Add a global variable to disable search autocomplete
The new search autocomplete is due to be rolled out across GOV.UK.

Autocomplete uses a denylist to stop certain terms from serving
suggestions. The process to update the denylist is very manual and
developer heavy.

If might be difficult to update the denylist over the holiday period.
In that situation the decision might be taken to temporarily turn off
autocomplete entirely.

A global environment variable has been added to each environment rather
than individual environment variables for each affect app, as the search
bar exists in the header on all pages on GOV.UK. This means that it
needs to be accessed by the layout_super_navigation_header [1] which is pulled in by
static for use in 6 other rendering apps.

We are using an env var to disable this feature rather than enable it so
that by default developers using these apps do not need to set-up extra
env vars to use the app as it runs in production, with an expectation
that it is unlikely that this feature would be disabled for a prolonged
period.

We use a conditional in the helm chart as a means to define whether an
env var is created or not - with the expectation that the env var with
any value is a indication the feature should be disabled. When
`disableSearchAutocomplete` is set to true in a
`values-<environment>.yaml` file the `GOVUK_DISABLE_SEARCH_AUTOCOMPLETE` will
be set and have a value of "1". When `disableSearchAutocomplete` is false then
the conditional will be false and no env var is set. Apps that use this env var
can do:

```
if ENV["GOVUK_DISABLE_SEARCH_AUTOCOMPLETE"]
  # render search without autocomplete
else
  # render something with autocomplete
end
```

It is expected that once the autocomplete feature is settled in GOV.UK
there will no longer be a need to disable it and this code can be
removed, ideally this will happen in Q4 2024-2025 while the Search team
is completing the deployment of this feature.

[1]: https://github.com/alphagov/govuk_publishing_components/pull/4451/files

Co-authored-by: Leena Gupte <leena.gupte@digital.cabinet-office.gov.uk>