Remove GOV.UK Chat resource increases
This commit is intended to be applied after the pilot of GOV.UK Chat
closes on the 23rd December.

This removes the customisations made to the govuk-chat helm chart to
increase the resources available to GOV.UK Chat in the event of a spike.
When the GOV.UK Chat pilot closes the application will still be running
but users viewing the frontend part will receive a 409 response. We do
however need to keep resources available as the app is still maintaining
a search index and still has an admin area that can be used for
analysis. Hence this just reduces the parts of our Kubernetes
infrastructure where we've pushed for additional resources beyond the
GOV.UK default.