Use the generated Signon API key in Whitehall
I’ve given the “Whitehall” user access to the newly-minted API
endpoint in Signon, and run the `signon-sync-token-secrets-to-k8s`
job, so this secret should be ready for use in each app.

Trying this in non-prod environments first. Will roll out to prod
if/when this is proved to be working.