Set Redis maxmemory config
The maxmemory and maxmemory-policy settings prevent Redis from caching
any further values if it reaches the limit. It will return an error to
the client. This prevents Redis from exceeding the container memory
limit and being OOM-killed. Because we have data persistence enabled,
when a new container is created, it immediately fills the memory from
the data on disk, causing a crash loop. This results in the
unavailability of Redis and prevents Sidekiq jobs from being processed.
Setting the memory limit will help keep Redis available for read
operations by Sidekiq workers.