Add Recreate Strategy for Redis Deployments
Redis uses a PVC with ReadWriteOnce, which causes a deadlock during pod
rollouts. With the rolling update strategy, a new pod is brought up
before the old pod is deleted. However, the new pod cannot start because
it waits for access to the PVC, which is still being used by the old
pod. Meanwhile, the old pod is not deleted until the new pod is created,
resulting in a deadlock.