De-integrate GOV.UK Chat from www.gov.uk in integration
Now that we have configured a gone route for www.gov.uk/chat [1] this
removes the additional configuration to serve the /chat route with the
www loadbalancer.

This removes the additional ALB conditions for router with extra paths
for /chat. It removes the nginx redirects to keep people on
www rather than the chat.publishing hostname, and turns on signon
authentication for chat again, so that it can't be accessed without a
signon account.

I've applied this only to integration so far so that I can verify it all
works correctly.

I've left the ingress configuration [2] as it seemed every other app
with public access has it, however I was a little confused as we didn't
seem to add it until we configured the app to be served on www.

[1]: https://github.com/alphagov/publishing-api/pull/3122
[2]: https://github.com/alphagov/govuk-helm-charts/commit/06f61038f801fbee1ad2c4790ee9fb81ac92e8e2