Add read replica for Publishing API in integration
This will be a parallel deployment of the Publishing API application in
intgegration, used solely for serving GraphQL queries from frontend
applications.

As this will not be receiving any write requests, it does not need
access to Redis or any workers.

The `hostReplica` secret has already been added through the AWS Console.
The other credentials remain the same as the main version of the
database.