Configure Publishing API read replica environment variables
The replica of this application needs access to both the primary and
replica databases, so providing both through environment varaibles.

The reason is that Publishing API has a `users` table which is written
to after authentication, and contains information such as the user's
permissions. The replica still needs to write to this table.

Additionally, it needs to have the `RAILS_ENV` overwritten, to tell the
application to operate in replica mode.