Remove Publishing API bearer tokens for GraphQL queries
Authentication for GraphQL requests in Publishing API was removed in
https://github.com/alphagov/publishing-api/pull/3220.

Therefore we no longer need a Publishing API bearer token in the
applications that make these queries.

Note: the token is being left in place for Government Frontend, since
this existed prior to the GraphQL work for another purpose (running a
rake task to publish static pages).