Create app-config-ephemeral chart, move most resources out of argo-bootstrap-ephemeral
We use the app-of-apps pattern for 'normal' clusters (app-config), but not for ephemeral clusters. This makes it difficult to change config for things managed by the bootstrap chart, since you have to make a new chart release, then update TF to point to the new version.

This PR brings the app-of-apps pattern to ephemeral clusters, so only a git merge is required to change config for these resources