Add global variables for new fact check service account
Gmail have officially deprecated and removed password based logins
for apps, instead making us work with OAuth2. As part of the migration
to this system we need new secrets available to Publisher for
fact checking.

This includes only integration for now for further testing of the
implementation before moving it to new environments.